encapsulates test into test package
