--- network/fixtures.go
@@ -1,47 +1,0 @@
-package network
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	message2 "github.com/onflow/flow-go/model/libp2p/message"
-	"github.com/onflow/flow-go/network/codec/json"
-)
-
-// NetworkPayloadFixture creates a blob of random bytes with the given size (in bytes) and returns it.
-// The primary goal of utilizing this helper function is to apply stress tests on the network layer by
-// sending large messages to transmit.
-func NetworkPayloadFixture(t *testing.T, size uint) []byte {
-	// reserves 1000 bytes for the message headers, encoding overhead, and libp2p message overhead.
-	overhead := 1000
-	require.Greater(t, int(size), overhead, "could not generate message below size threshold")
-	emptyEvent := &message2.TestMessage{
-		Text: "",
-	}
-
-	// encodes the message
-	codec := json.NewCodec()
-	empty, err := codec.Encode(emptyEvent)
-	require.NoError(t, err)
-
-	// max possible payload size
-	payloadSize := int(size) - overhead - len(empty)
-	payload := make([]byte, payloadSize)
-
-	// populates payload with random bytes
-	for i := range payload {
-		payload[i] = 'a' // a utf-8 char that translates to 1-byte when converted to a string
-	}
-
-	event := emptyEvent
-	event.Text = string(payload)
-	// encode event the way the network would encode it to get the size of the message
-	// just to do the size check
-	encodedEvent, err := codec.Encode(event)
-	require.NoError(t, err)
-
-	require.InDelta(t, len(encodedEvent), int(size), float64(overhead))
-
-	return payload
-}
