+++ network/protocol/readConnection.go
