+++ network/protocol/readSubscription.go
